refactor(check): extract match.ts from checker.ts
Move pattern matching functionality to a dedicated module:
- Match arm detection and processing
- Pattern validation (literal, or, wildcard, binding)
- Catch-all pattern detection
- Match exhaustiveness checking
- Match finalization and binding creation

This is Layer 3 (Constructs) in the checker module hierarchy.
